# openapi.yaml
openapi: 3.0.1
info:
  title: Ledger API
  description: Account ledger guinea pig for TrueCD pipeline validation
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development
paths:
  /accounts:
    post:
      tags:
        - Accounts
      summary: Create a new account
      operationId: createAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /accounts/{id}:
    get:
      tags:
        - Accounts
      summary: Get account by UUID
      operationId: getAccountById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Account found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /accounts/number/{accountNumber}:
    get:
      tags:
        - Accounts
      summary: Get account by account number
      operationId: getAccountByNumber
      parameters:
        - name: accountNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Account found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /accounts/{id}/freeze:
    patch:
      tags:
        - Accounts
      summary: Freeze an account
      operationId: freezeAccount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Account frozen
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Account not active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /accounts/{id}/close:
    patch:
      tags:
        - Accounts
      summary: Close an account
      operationId: closeAccount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Account closed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Account not active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /accounts/{id}/transactions:
    post:
      tags:
        - Accounts
      summary: Post a credit or debit transaction
      operationId: postTransaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostTransactionRequest'
      responses:
        '201':
          description: Transaction posted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Account not active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Insufficient funds
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - Accounts
      summary: List all transactions (newest first)
      operationId: listTransactions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Transaction list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionResponse'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /accounts/{id}/balance:
    get:
      tags:
        - Accounts
      summary: Get current balance (Redis-cached)
      operationId: getBalance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Balance response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /actuator/health:
    get:
      tags:
        - Actuator
      summary: Health check
      responses:
        '200':
          description: Application health status

  /actuator/health/readiness:
    get:
      tags:
        - Actuator
      summary: Readiness probe
      responses:
        '200':
          description: Application readiness status

components:
  schemas:
    CreateAccountRequest:
      type: object
      required:
        - ownerName
      properties:
        ownerName:
          type: string
          maxLength: 100
          example: Juan Dela Cruz
        currency:
          type: string
          pattern: '[A-Z]{3}'
          default: PHP
          example: PHP

    AccountResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        accountNumber:
          type: string
          example: ACC1234567
        ownerName:
          type: string
        currency:
          type: string
        status:
          type: string
          enum: [ACTIVE, FROZEN, CLOSED]
        createdAt:
          type: string
          format: date-time
        currentBalance:
          type: number
          format: decimal

    PostTransactionRequest:
      type: object
      required:
        - type
        - amount
      properties:
        type:
          type: string
          enum: [CREDIT, DEBIT]
        amount:
          type: number
          format: decimal
          minimum: 0.01
        description:
          type: string

    TransactionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        accountId:
          type: string
          format: uuid
        type:
          type: string
          enum: [CREDIT, DEBIT]
        amount:
          type: number
          format: decimal
        description:
          type: string
        balanceAfter:
          type: number
          format: decimal
        createdAt:
          type: string
          format: date-time

    BalanceResponse:
      type: object
      properties:
        accountId:
          type: string
          format: uuid
        accountNumber:
          type: string
        balance:
          type: number
          format: decimal
        currency:
          type: string
        cachedAt:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time